requirements.txt
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
faster-whisper>=1.0.0
rumps>=0.4.0
pyannote.audio>=3.1.0
